+++ docs/changes.rst
@@ -12,6 +12,7 @@ Released on ? 2016.
 * Project owners can now configure translation component and project details.
 * Use "Watched" instead of "Subscribed" projects.
 * Projects can be watched directly from project page.
+* Added multi language status widget.
 
 weblate 2.8
 -----------